Fix config loading/saving.
The config subsystem assumes we can go between an `int` and `char *`
without issue.  This is an incorrect assumption, so fix the config
loading to encode the expected type and use a union to reference the
proper config value.

Of note, the previous string loading was completely broken for some
values.  You need to store a reference to the char * for it to work
properly.